refactor: DTOs usando records.

+++ src/main/java/apolinario0x21/toDoApp/service/TaskService.java
@@ -1,9 +1,11 @@
 package apolinario0x21.toDoApp.service;
 
-import apolinario0x21.toDoApp.dto.TaskRequest;
+import apolinario0x21.toDoApp.dto.TaskRequestDTO;
+import apolinario0x21.toDoApp.dto.TaskResponseDTO;
 import apolinario0x21.toDoApp.exceptions.TaskNotFoundException;
 import apolinario0x21.toDoApp.model.Task;
 import apolinario0x21.toDoApp.repository.TaskRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,38 +14,63 @@ import java.util.UUID;
 
 @Service
 public class TaskService {
+
     @Autowired
     private TaskRepository taskRepository;
 
-    public List<Task> getAllTasks() {
-        return taskRepository.findAllByOrderByCreatedAtDesc();
+    public TaskResponseDTO toResponse(Task task) {
+
+        return new TaskResponseDTO(
+                task.getId(),
+                task.getTitle(),
+                task.isCompleted(),
+                task.getCreatedAt());
     }
 
-    public Task createTask(TaskRequest request) {
+    public TaskResponseDTO createTask(TaskRequestDTO request) {
+
         Task task = new Task();
-        task.setTitle(request.getTitle());
-        return taskRepository.save(task);
+        task.setTitle(request.title());
+        task.setCompleted(false);
+        Task savedTask = taskRepository.save(task);
+        return toResponse(savedTask);
     }
 
-    public TaskRequest getTaskById(UUID id) {return null;}
+    public List<TaskResponseDTO> findAllTasks() {
+        return taskRepository.
+                findAllByOrderByCreatedAtDesc()
+                .stream().map(this::toResponse).toList();
+    }
+    //getTaskById(UUID id)
 
-    public void updateTaskTitle(UUID id, String newTitle) {
-        Task task = taskRepository.findById(id)
+    private Task findById(UUID id) {
+        return taskRepository.findById(id)
                 .orElseThrow(() -> new TaskNotFoundException("Task not found. ID:" + id));
+    }
+
+    public TaskResponseDTO findTaskById(UUID id) {
+        Task task = findById(id);
+        return toResponse(task);
+    }
+
+    public void titleUpdateTask(UUID id, String newTitle) {
+        Task task = findById(id);
         task.setTitle(newTitle);
         taskRepository.save(task);
     }
 
-    public void updateTaskStatus(UUID id, boolean completed) {
-        Task task = taskRepository.findById(id)
-                .orElseThrow(() -> new TaskNotFoundException("Status not found"));
+    public void statusUpdateTask(UUID id, boolean completed) {
+        Task task = findById(id);
         task.setCompleted(completed);
         taskRepository.save(task);
     }
 
     public void deleteTask(UUID id) {
-        taskRepository.deleteById(id); // deleta a tarefa com o id especificado
-        // id/elemento existe?
+        if (!taskRepository.existsById(id)) {
+            throw new TaskNotFoundException("Task not found. ID:" + id);
+        }
+
+        taskRepository.deleteById(id);
     }
